lua: initial Lua bits of multi-skill execution
Implemented first rough version of the Lua code required for multi-skill
execution. There is still more cleanup necessary, but it's a start and
with the coming skiller changes in principal able to execute and monitor
multiple skills. Dependency code missing, turns out to be crucial in
multi-skill environment to be able to properly detect if for example
user forgot to call stop before executing the same skill string again.

Implemented SkillChannel class which now takes over handling of the
executiong and monitoring of a single skill string. The new
skiller.exec module binds this together to have multiple channels.
skiller.skillenv lost several functions, and will probably loose even
more in the near future.

Conflicts:
	src/lua/skiller/skillenv.lua
	src/lua/skiller/start.lua