colli: renamed AStar class to avoid conflict with utils/search/astar
Since we compiled the utils with C++11, the linker mixes up the colli
astar and utils astar classes. They are implemented differently, but
have the same name and namespace. This lead to a segfault when suddenly
the search function of the utils was executed in a colli astar instance.
This is a fist hotfix that just renames the colli astar class.