robot-memory: fixed initial creation of dbs to be able to move them
DBs have to be created before they can be moved to a primary. To be
created they have to be not empty. The previous initialization created
the db in the local mongod but the mongos pointed to an empty  db
with the same name in another mongod. Thus it produced the error
'can't find db'. Now the db is created using the mongos and afterwards
moved to the local mongod.