clips: retract blackboard facts only when updating
Blackboard facts were cleaned up by a lowest-priority rule after each
evaluation cyle. This meant that blackboard facts are gone for good if
there isn't some rule that immediately fires on them. This effectively
prevents the agent from watching for blackboard conditions whenever
there are other conditions involved.

This commit changes the behaviour so that blackboard facts are retracted
immediately before they are updated, so the fact will exist as long as
it remains valid.

I'm not sure whether this behaviour even needs to be made configurable
since I find it highly unlikely that it will break existing assumptions.
Will continue testing with the new behaviour and make it switchable if
problems turn up after all.