libinterface: copy full type ID and then enforce null-terminated strings
The functions strncpy and snprintf expect the null terminator to be
within the specified size of the string:
  The  strncpy() function is similar, except that at most n bytes of src
  are copied.  Warning: If there is no null byte among the first n bytes
  of src, the string placed in dest will not be null-terminated.

  The functions snprintf() and vsnprintf() write at most size bytes
  (including the terminating  null  byte ('\0')) to str.

GCC7 complains if the input string of snprintf() is longer than the
destination string. For consistency reasons, always strncpy() and
snprintf() the whole string (including the last byte), and then enforce
the string to be null-terminated. This retains the old behavior
(truncating the string if it isn't null-terminated) but avoids the gcc
error.