stn-generator: initial commit with generator copied over
The generator was first implemented and tested externally with a toy
domain and plan. The infrastructure is now copied into this plugin to
use it within Fawkes.

The plugin will import a sequential plan and a domain from robot-memory
and generate a Simple Temporal Network which is again exported into
robot-memory.