clips: add PDDL feature
This adds a function parse-pddl-domain to CLIPS. The function will call
the PDDL parser on the given domain file and assert domain facts
according to the given domain. The parser currently only asserts facts
for the types, operators, and operator preconditions of a domain.
Effects and constants are to be done.