clips-executive: Fix, satisfiability check of conjunct precond
Fixed-case:
If the atomic-preconditions/nested-precondition where never grounded,
the satisfiability check of the parent precond will still hold making
the conjunction satisfiable and hence the action executable
(EVEN though there is not grounded atomic-preconditions)

This is a temporary fix till A General-Grounding-Function is
implemented. That will solve this issue as a by product.